+++ fiddles/extjs6/fiddle-0000-Template/src/meta.js
@@ -1,0 +1,9 @@
+var meta = {
+    fiddleHeader: 'fiddle-0000-Template',
+    fiddleSubHeader: 'Template Fiddle created @ 09-30-16' +
+    '<br />'
+};
+
+var fiddleStore = null;
+
+