+++ fiddles/extjs6/fiddle-0036-ItunesTop100/src/meta.js
@@ -1,0 +1,9 @@
+var meta = {
+    fiddleHeader: 'fiddle-0036-ItunesTop100',
+    fiddleSubHeader: 'Template Fiddle created @ 09-29-16' +
+    '<br />'
+};
+
+var fiddleStore = null;
+
+